refactor(KERN): run lint
