+++ app/components/kern/formElements/exclusiveCheckboxes/KernExclusiveCheckboxInput.tsx
@@ -1,5 +1,5 @@
 import classNames from "classnames";
-import { type CheckboxValue } from "~/components/formElements/Checkbox";
+import { type CheckboxValue } from "~/components/formElements/KernCheckbox";
 import { useJsAvailable } from "~/components/hooks/useJsAvailable";
 
 type KernExclusiveCheckboxInputProps = {
